refactor: 주문 계산 로직 useOrderCalculation 훅으로 분리
