+++ src/hooks/order/OrderConstants.ts
@@ -1,0 +1,2 @@
+export const FREE_SHIPPING_STANDARD = 100000;
+export const SHIPPING_FEE = 3000;
